pyproject.toml
[project]
name = "agent-skills-mcp"
version = "0.1.0"
description = "Load agent skills as MCP tools"
readme = "README.md"
requires-python = ">=3.13"
authors = [
    {name = "DiscreteTom", email = "discrete_tom@outlook.com"},
]
license = {text = "MIT"}
keywords = ["mcp", "tools", "model-context-protocol", "agent-skills"]
classifiers = [
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
dependencies = [
    "fastmcp>=2.12.4",
    "python-frontmatter>=1.1.0",
]

[project.urls]
Homepage = "https://github.com/DiscreteTom/agent-skills-mcp"
Repository = "https://github.com/DiscreteTom/agent-skills-mcp"
Issues = "https://github.com/DiscreteTom/agent-skills-mcp/issues"

[project.scripts]
agent-skills-mcp = "agent_skills_mcp.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.sdist]
exclude = [
  "/.pre-commit-config.yaml",
  "/.python-version", 
  "/pyrightconfig.json",
  "/uv.lock"
]

[dependency-groups]
dev = [
    "black>=25.9.0",
    "hatchling>=1.27.0",
    "pre-commit>=4.3.0",
    "pyright>=1.1.405",
    "ruff>=0.13.2",
    "pytest>=8.0.0",
    "pytest-cov>=6.0.0",
    "pytest-asyncio>=1.2.0",
]
